requirements.txt
Pillow==9.5.0
qrcode==7.4.2
streamlit==1.22.0
streamlit_lottie==0.0.3
